backend/requirements.txt
# MSTool-AI-QMS — Regulatory Compliance Automation Platform
# FastAPI Backend

# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2

# HTTP Client (for GitHub API, NVD API, MSTool-AI API)
httpx==0.28.1

# MCP Servers
fastmcp==2.3.0

# AI (Claude API for agents)
anthropic==0.44.0

# Firebase Admin SDK
firebase-admin>=6.5.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT>=2.10.1

# Utilities
python-dotenv>=1.1.0
python-json-logger==2.0.7
aiofiles==24.1.0

# In-process scheduler — runs hourly snapshots + daily sentinel inside Cloud Run
# (avoids the need to enable Cloud Scheduler API on the GCP project)
APScheduler==3.10.4

# Rate limiting on AI agent invocations (Claude API cost protection)
slowapi==0.1.9

# Excel export for the CE Mark Submission Package (RTM matrix)
openpyxl==3.1.5

# Cryptographic signatures (21 CFR Part 11 §11.50 PKCS#7 detached sigs).
# Cloud KMS does the actual signing; cryptography reassembles the PKCS#7
# envelope locally so the verification path needs no GCP roundtrip.
cryptography>=42.0.0
google-cloud-kms>=3.0.0

# PDF Generation
reportlab==4.2.5

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov
